Change the attachment of the object instances into a scene
When an object instance is attached to a scene, it is added to a hash
table that maps the Star-3D identifier of the object instance to the
object instance itself. Contrary to the previous linked list this
associative container allows to retrieve the object instance directly
from its Star-3D identifier.