Add a Star-3D sampling shape to the ssol_shape.
Use the Star-3D sampling shape to create a Star-3D sampling scene in the
ssol_object. Instantiate it as a sampling instance in the
ssol_object_instance. On solve invocation, conditionally attach the
Star-3D sampling instance to the Star-3D sampling scene of the
ssol_scene whether its associated material is a mirror or not. Fill an
associative container that maps the Star-3D instance identifier to its
associated ssol_object_instance and use it to retrieve the sampled
solstice instance at the beginning of each realisation.

Clean up the code by removing useless getters.