Fix the ssol_solve function
The Monte-Carlo estimation of the "missing" radiative flux took into
account the shadowed radiative flux, i.e. the part of the radiative flux
that does not reach a sampled primitive.

Furthermore, a valid radiative path that did not hit anything was not
handled into the "missing" MC estimation due to a bad branching.