Fix the integration with an atmosphere having a spectral absorption
The absorption of the surrounding medium was assumed to be a scalar.
However the atmosphere can have a spectral absorption, which was not
consistent with the previous assumption leading to errors when checking
for media consistency. T

This commit setups the surrounding medium with the absorption of the
atmosphere. If no atmosphere is defined the absorption is null.