Fix when the sampled starting point lies into a dielectric material
The first vertex of a radiative path is assumed to lie in the
surrounding medium. This may be actually wrong if the starting point is
sampled on a geometry with a dielectric material. Indeed, the side of
the sampled face is defined in order to ensure that the sampled point
looks toward the sun and thus the starting point can lie on the internal
side of the dielectric geometry. In this case, this point lies in the
internal medium of the geometry and not in the surrounding medium.

This commit ensures that the starting medium is the correct one in order
to ensure that the media consistency tests do not fail.