Fix the upd of path position when it restarts from a virtual surface
When a ray hits a surface with a virtual material, some treatments are
performed and it pursues its trip from the intersection point. In order
to avoid numerical inaccuracy, the origin of the new "departure" is not
move to the hit position. Instead, the minimum boundary of the original
ray is set just after the previous hit distance and the new ray is
traced with the same origin.

This algorithmic trick was not correctly handled in the update of the
radiative path position. This commit fixes this issue.